requirements.txt
pandas
python-dotenv
ipykernel
plotly
loguru
click
boto3